updated the code of the bankSlider and the loading info faster

+++ lib/Constants/font_manager.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-//import 'package:flutter/widgets.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class FontManager {
@@ -23,12 +22,12 @@ class FontManager {
     String? fontFamily,
   }) {
     final style = GoogleFonts.roboto(
-      fontSize: fontSize,
+      fontSize: fontSize.toDouble(),
       fontWeight: lWeight,
-      height: lineHeight,
+      height: lineHeight.toDouble(),
       fontStyle: lFontStyle,
       color: color,
-      letterSpacing: letterSpacing,
+      letterSpacing: letterSpacing.toDouble(),
       decoration: decoration,
       decorationColor: decorationColor,
       decorationThickness: decorationThickness,
@@ -38,55 +37,7 @@ class FontManager {
     return fontFamily == null ? style : style.copyWith(fontFamily: fontFamily);
   }
 }
-// import 'package:flutter/material.dart';
-// import 'package:google_fonts/google_fonts.dart';
 
-// class FontManager {
-//   static const double _baseWidth = 375; // Figma design width
-
-//   double _scale(BuildContext context, double size) {
-//     final screenWidth = MediaQuery.of(context).size.width;
-
-//     double scaled = size * (screenWidth / _baseWidth);
-
-//     // Clamp → prevents text from becoming too big/small
-//     return scaled.clamp(size * 0.9, size * 1.3);
-
-//   }
-
-//   TextStyle getTextStyle(
-//     BuildContext context, {
-//     Color color = Colors.black,
-//     FontWeight lWeight = FontWeight.normal,
-//     double lineHeight = 1.0,
-//     FontStyle lFontStyle = FontStyle.normal,
-//     double fontSize = 18.0,
-//     double letterSpacing = 0.24,
-//     TextDecoration decoration = TextDecoration.none,
-//     Color decorationColor = Colors.black,
-//     double decorationThickness = 0.0,
-//     TextDecorationStyle decorationStyle = TextDecorationStyle.solid,
-//     TextOverflow overflow = TextOverflow.visible,
-//     int? maxLines,
-//     TextAlign textAlign = TextAlign.start,
-//     TextDirection textDirection = TextDirection.ltr,
-//     bool softWrap = false,
-//   }) {
-//     return GoogleFonts.roboto(
-//       fontSize: _scale(context, fontSize), // ✅ automatic scaling
-//       fontWeight: lWeight,
-//       height: lineHeight,
-//       fontStyle: lFontStyle,
-//       color: color,
-//       letterSpacing: letterSpacing,
-//       decoration: decoration,
-//       decorationColor: decorationColor,
-//       decorationThickness: decorationThickness,
-//       decorationStyle: decorationStyle,
-
-//     );
-//   }
-// }
 class FontManager2 {
   TextStyle getTextStyle(
     BuildContext context, {
